print order

+++ db/migrate/20230516162959_create_orders.rb
@@ -2,11 +2,11 @@ class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
       t.references :client, null: false, foreign_key: true
-      t.decimal :subtotal
-      t.decimal :gst
-      t.decimal :pst
-      t.decimal :discount
-      t.decimal :total
+      t.float :subtotal
+      t.float :gst
+      t.float :pst
+      t.float :discount
+      t.float :total
       t.string :notes
       t.datetime :deadline
 
@@ -19,7 +19,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
       t.bigint :order_id
       t.bigint :product_id
 
-      t.decimal :subtotal
+      t.float :subtotal
       t.integer :quantity
     end
