+++ db/migrate/20230522132028_change_order_items_quantity_type.rb
@@ -1,5 +1,5 @@
 class ChangeOrderItemsQuantityType < ActiveRecord::Migration[7.0]
   def change
-    change_column(:order_items, :quantity, :decimal)
+    change_column(:order_items, :quantity, :float)
   end
 end
