+++ db/migrate/20230531204440_create_order_items.rb
@@ -2,8 +2,8 @@ class CreateOrderItems < ActiveRecord::Migration[7.0]
   def change
     drop_table :order_items
     create_table :order_items do |t|
-      t.decimal :quantity
-      t.decimal :subtotal
+      t.float :quantity
+      t.float :subtotal
       t.references :product, null: false, foreign_key: true
       t.references :order, null: false, foreign_key: true
